+++ lib/main_db/core/daos/history_dao/document_history_dao.dart
@@ -172,7 +172,6 @@ class DocumentHistoryDao extends DatabaseAccessor<MainStore>
         action: ActionInHistoryX.fromString(dto.action),
         name: dto.title,
         description: Value(dto.description),
-        categoryName: Value(dto.categoryName),
         usedCount: Value(dto.usedCount),
         isFavorite: Value(dto.isFavorite),
         isArchived: Value(dto.isArchived),